Create MInHeap.cpp
A Min Heap Binary Tree is a Binary Tree where the root node has the minimum key in the tree.

The above definition holds true for all sub-trees in the tree. This is called the Min Heap property.

Almost every node other than the last two layers must have two children. That is, this is almost a complete binary tree, with the exception of the last 2 layer.